refactor: game 패키지 생성 후 BoardCreator, InitialBoardCreator, InitialPosition 이동
