+++ src/main/java/chess/domain/game/GameCreator.java
@@ -1,10 +1,11 @@
-package chess.domain.board;
+package chess.domain.game;
 
+import chess.domain.board.BoardStatus;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import java.util.Map;
 
-public interface BoardCreator {
+public interface GameCreator {
 
     Map<Position, Piece> createBoard();
