+++ src/test/java/chess/domain/game/GameTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import static chess.fixture.PieceFixture.BLACK_KING;
 import static chess.fixture.PieceFixture.WHITE_QUEEN;
@@ -14,9 +14,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.board.BoardCreator;
+import chess.domain.game.GameCreator;
 import chess.domain.board.BoardStatus;
-import chess.domain.board.InitialBoardCreator;
+import chess.domain.game.InitialGameCreator;
+import chess.domain.game.Game;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import java.util.HashMap;
@@ -41,7 +42,7 @@ class GameTest {
     @DisplayName("검은색 진영이 먼저 턴을 진행하면 예외가 발생한다.")
     @Test
     void notProceedBlackTurn() {
-        Game game = new Game(new InitialBoardCreator());
+        Game game = new Game(new InitialGameCreator());
 
         assertThatThrownBy(() -> game.proceedTurn(A7, A6))
                 .isInstanceOf(IllegalArgumentException.class)
@@ -51,7 +52,7 @@ class GameTest {
     @DisplayName("흰색 진영이 먼저 턴을 진행한다.")
     @Test
     void proceedWhiteTurn() {
-        Game game = new Game(new InitialBoardCreator());
+        Game game = new Game(new InitialGameCreator());
 
         assertThatCode(() -> game.proceedTurn(A2, A3))
                 .doesNotThrowAnyException();
@@ -72,7 +73,7 @@ class GameTest {
     @DisplayName("흰색 진영 다음에 흰색 진영이 턴을 진행하면 예외가 발생한다.")
     @Test
     void notProceedWhiteTurn() {
-        Game game = new Game(new InitialBoardCreator());
+        Game game = new Game(new InitialGameCreator());
         game.proceedTurn(A2, A3);
 
         assertThatThrownBy(() -> game.proceedTurn(B2, B3))
@@ -83,7 +84,7 @@ class GameTest {
     @DisplayName("흰색 진영 다음에 검은색 진영이 턴을 진행한다.")
     @Test
     void proceedBlackTurn() {
-        Game game = new Game(new InitialBoardCreator());
+        Game game = new Game(new InitialGameCreator());
         game.proceedTurn(A2, A3);
 
         assertThatCode(() -> game.proceedTurn(A7, A6))
@@ -105,7 +106,7 @@ class GameTest {
     @DisplayName("King을 잡다.")
     @Test
     void catchKing() {
-        Game game = new Game(new BoardCreator() {
+        Game game = new Game(new GameCreator() {
             @Override
             public Map<Position, Piece> createBoard() {
                 return new HashMap<>() {{
